Added possibility to run script with LinuxMint
I added an method to prob for derivatives. If a derivative is found (Linux Mint)  lsb_dist_id, lsb_release and lsb_codename are set for the mother distribution (Ubuntu).
The user gets an dialog that this is an not official supported derivative and could check the selected mother distribution.